requirements.txt
joblib==1.0.1
numpy==1.20.1
pandas==1.2.4
scikit-learn==0.24.2
streamlit==0.88.0
Pillow~=8.3.2